refactor: prediction classes to own directory
