+++ src/dotnet-affected/Infrastructure/Prediction/ChangedProjectsProvider.cs
@@ -1,10 +1,8 @@
 ﻿using Affected.Cli.Commands;
-using Microsoft.Build.Execution;
 using Microsoft.Build.Graph;
 using Microsoft.Build.Prediction;
 using Microsoft.Build.Prediction.Predictors;
 using Microsoft.Build.Prediction.Predictors.CopyTask;
-using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -87,97 +85,6 @@ namespace Affected.Cli
                         yield return key;
                     }
                 }
-            }
-        }
-
-        private class ProjectInputFilesCollector : IProjectPredictionCollector
-        {
-            private readonly string _repositoryPath;
-            public HashSet<string> AllFiles { get; } = new HashSet<string>();
-
-            public ProjectInputFilesCollector(string repositoryPath)
-            {
-                _repositoryPath = repositoryPath;
-            }
-
-            public void AddInputFile(string path, ProjectInstance projectInstance, string predictorName)
-            {
-                // Make the path absolute if needed.
-                if (!Path.IsPathRooted(path))
-                {
-                    path = Path.GetFullPath(Path.Combine(projectInstance.Directory, path));
-                }
-                else if (!path.StartsWith(_repositoryPath))
-                {
-                    // ignore files outside the project's directory
-                    return;
-                }
-
-                this.AllFiles.Add(path);
-            }
-
-            public void AddInputDirectory(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-
-            public void AddOutputFile(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-
-            public void AddOutputDirectory(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-        }
-
-        private class FilesByProjectGraphCollector : IProjectPredictionCollector
-        {
-            private readonly Dictionary<ProjectInstance, ProjectInputFilesCollector> _collectorByProjectInstance;
-
-            public Dictionary<ProjectGraphNode, ICollection<string>> PredictionsPerNode { get; }
-
-            public FilesByProjectGraphCollector(ProjectGraph projectGraph, string repositoryPath)
-            {
-                var projectGraphNodes = projectGraph.ProjectNodes;
-
-                PredictionsPerNode = new Dictionary<ProjectGraphNode, ICollection<string>>(projectGraphNodes.Count);
-
-                _collectorByProjectInstance =
-                    new Dictionary<ProjectInstance, ProjectInputFilesCollector>(projectGraphNodes.Count);
-
-                foreach (var projectGraphNode in projectGraphNodes)
-                {
-                    var collector = new ProjectInputFilesCollector(repositoryPath);
-                    PredictionsPerNode.Add(projectGraphNode, collector.AllFiles);
-                    _collectorByProjectInstance.Add(projectGraphNode.ProjectInstance, collector);
-                }
-            }
-
-            public void AddInputFile(string path, ProjectInstance projectInstance, string predictorName) =>
-                GetProjectCollector(projectInstance)
-                    .AddInputFile(path, projectInstance, predictorName);
-
-            public void AddInputDirectory(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-
-            public void AddOutputFile(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-
-            public void AddOutputDirectory(string path, ProjectInstance projectInstance, string predictorName)
-            {
-            }
-
-            private ProjectInputFilesCollector GetProjectCollector(ProjectInstance projectInstance)
-            {
-                if (!_collectorByProjectInstance.TryGetValue(projectInstance,
-                        out ProjectInputFilesCollector? collector))
-                {
-                    throw new InvalidOperationException(
-                        "Prediction collected for ProjectInstance not in the ProjectGraph");
-                }
-
-                return collector;
             }
         }
     }
